pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app-manifest-cli"
version = "0.1.0"
description = "Generate Application Manifest v2 (CycloneDX 1.6) JSON files"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "jsonschema>=4.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app_manifest"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[project.scripts]
am = "app_manifest.cli:cli"
